% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_from_function-function.R
\name{edge_from_function}
\alias{edge_from_function}
\title{Identify edges given functions}
\usage{
edge_from_function(func)
}
\arguments{
\item{func}{Functions, an object class generated by
\code{\link{function_from_edge}} or \code{\link{function_from_user}}
functions.}
}
\value{
A data frame which include the columns 'from' and 'to in this order.
}
\description{
Identify edges given functions
}
\examples{

data(functions)
edge_from_function(functions)
}
\keyword{function-specified-edge-listing}
