% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges-data.R
\docType{data}
\name{edges}
\alias{edges}
\title{Edges}
\format{
A data frame with 7 rows and 2 columns:
\describe{
\item{from}{A vertex name from which a directed edge comes.}
\item{to}{A vertex name to which a directed edge comes.}
}
}
\source{
Generated for examples in this package.
}
\usage{
edges
}
\description{
An example of a data frame which include the columns 'from' and 'to' in this
order. A vertex name 'n' does not exist.
}
\keyword{dataset}
