% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_from_edge-function.R
\name{function_from_edge}
\alias{function_from_edge}
\title{List functions given edges}
\usage{
function_from_edge(e)
}
\arguments{
\item{e}{Edge, a data frame that must only include the columns 'from'
and 'to in this order. A vertex name 'n' is not allowed.}
}
\value{
A list of character vectors of arguments for function which will be
defined by a user using \code{\link{define}} function.
}
\description{
List functions given edges
}
\examples{

data(edges)
function_from_edge(edges)
}
\keyword{edge-specified-function-listing}
