% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-data.R
\docType{data}
\name{functions}
\alias{functions}
\title{Functions}
\format{
A list with 5 elements:
\describe{
\item{B}{A function with an argument 'n'.}
\item{A}{A function with an argument 'B'.}
\item{D}{A function with an argument 'A'.}
\item{C}{A function with arguments 'A', 'B', and 'D'.}
\item{E}{A function with arguments 'A' and 'C'.}
}
}
\source{
Generated for examples in this package.
}
\usage{
functions
}
\description{
An example of an object class generated by \code{function_from_edge}
or \code{function_from_user} functions. The causal structure is a directed
acyclic graph (DAG), which means no loops are allowed. A function in the list
include 'n' as the only argument. All arguments within any function are
defined by their respective functions, except the argument 'n'. The output
lengths of vertex functions match the specified length 'n'.
}
\keyword{dataset}
