% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Functions-function.R
\name{print.Functions}
\alias{print.Functions}
\title{Print method for Functions}
\usage{
\method{print}{Functions}(x, ...)
}
\arguments{
\item{x}{Functions, an object class generated by
\code{\link{function_from_edge}} or \code{function_from_user} functions.}

\item{...}{Additional arguments are ignored in this method, but are included
to maintain consistency with the generic print method.}
}
\value{
A summary of vertices that has functions. If there are vertices
without functions, an instruction is shown.
}
\description{
Print method for Functions
}
\examples{

data(edges)
functions <- function_from_edge(edges)
print(functions)
}
\keyword{function}
